Move fixture test data to test_fixture
